Initiate move to ASSYST
Start with structures; split the individual steps into their own
modules and added a job (untested)
