+++ .ci_support/environment.yml
@@ -11,6 +11,7 @@ dependencies:
 - pyiron_atomistics =0.6.4
 - pyiron-data =0.0.25
 - pyiron_snippets =0.1.1
+- pyiron_contrib >0.1.16
 - pyparsing =3.1.1
 - scipy =1.14.0
 - dill >=0.3
