+++ pyiron_potentialfit/assyst/util.py
