+++ setup.py
@@ -39,6 +39,8 @@ setup(
         'runnerase==0.3.3',
         'pyiron_snippets==0.1.1',
         # spgfit
+        'pyiron_atomistics==0.6.4',
+        'pyiron_contrib==0.1.16',
         'dill>=0.3.0',
         'seaborn>=0.13.0,<0.14',
         'pyxtal>=0.6.0,<0.7',
